acquire.py
